requirements.txt
pyqt6
natsort
python-dotenv
wheel
git+https://github.com/kramcat/CharacterAI.git
qasync
nltk
torch
translate

pyopenjtalk
espnet==202308
pypinyin==0.44.0
parallel_wavegan==0.5.4
gdown==4.4.0
espnet_model_zoo
sounddevice

tqdm